pyproject.toml
[project]
name = "adversarial-patch-vlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adversarial-robustness-toolbox>=1.20.1",
    "clip",
    "ftfy>=6.3.1",
    "kaggle>=2.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "regex>=2026.4.4",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "transformers>=5.5.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
